Revert "Arreglo de compatibilidades"
