--- src/main/java/com/unla/grupo21/entities/Persona.java
@@ -1,145 +1,0 @@
-package com.unla.grupo21.entities;
-
-import java.time.LocalDateTime;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Inheritance;
-import javax.persistence.InheritanceType;
-import javax.persistence.Table;
-
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-import org.springframework.data.annotation.ReadOnlyProperty;
-
-import com.unla.grupo21.models.TipoDocumento;
-
-@Entity
-@Table(name="persona")
-@Inheritance(strategy = InheritanceType.JOINED)
-public class Persona {
-
-	@Id
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-	
-	@Column(name="nombre", nullable= false)
-	private String nombre;
-	
-	@Column(name="apellido", nullable= false)
-	private String apellido;
-	
-	@Column(name="tipodocumento", nullable= false)
-	@Enumerated(EnumType.ORDINAL)
-	private TipoDocumento tipoDocumento;
-	
-	@Column(name="documento", nullable= false)
-	private long documento;
-	
-	@Column(name="email", nullable= false)
-	private String email;
-	
-	@Column(name="createdat", nullable = false, updatable = false)
-	@CreationTimestamp
-	@ReadOnlyProperty
-	private LocalDateTime createdAt;
-	
-	@Column(name="updatedat")
-	@UpdateTimestamp
-	private LocalDateTime updatedAt;
-	
-	public Persona() {}
-
-	public Persona(String nombre, String apellido, TipoDocumento tipoDocumento, long documento, String email) {
-		super();
-		this.nombre = nombre;
-		this.apellido = apellido;
-		this.tipoDocumento = tipoDocumento;
-		this.documento = documento;
-		this.email = email;
-	};
-	
-	public Persona(int id, String nombre, String apellido, TipoDocumento tipoDocumento, long documento, String email) {
-		super();
-		this.id = id;
-		this.nombre = nombre;
-		this.apellido = apellido;
-		this.tipoDocumento = tipoDocumento;
-		this.documento = documento;
-		this.email = email;
-	}
-
-	public int getId() {
-		return id;
-	}
-
-	public void setId(int id) {
-		this.id = id;
-	}
-
-	public String getNombre() {
-		return nombre;
-	}
-
-	public void setNombre(String nombre) {
-		this.nombre = nombre;
-	}
-
-	public String getApellido() {
-		return apellido;
-	}
-
-	public void setApellido(String apellido) {
-		this.apellido = apellido;
-	}
-
-	public TipoDocumento getTipoDocumento() {
-		return tipoDocumento;
-	}
-
-	public void setTipoDocumento(TipoDocumento tipoDocumento) {
-		this.tipoDocumento = tipoDocumento;
-	}
-
-	public long getDocumento() {
-		return documento;
-	}
-
-	public void setDocumento(long documento) {
-		this.documento = documento;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public LocalDateTime getCreatedAt() {
-		return createdAt;
-	}
-
-	public void setCreatedAt(LocalDateTime createdAt) {
-		this.createdAt = createdAt;
-	}
-
-	public LocalDateTime getUpdatedAt() {
-		return updatedAt;
-	}
-
-	public void setUpdatedAt(LocalDateTime updatedAt) {
-		this.updatedAt = updatedAt;
-	};
-	
-	
-	
-	
-	
-}
