+++ src/main/java/com/unla/grupo21/entities/User.java
@@ -1,17 +1,47 @@
 package com.unla.grupo21.entities;
 
+import java.time.LocalDateTime;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import javax.persistence.ManyToOne;
 import javax.persistence.Table;
+
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.annotation.ReadOnlyProperty;
 
 import com.unla.grupo21.models.TipoDocumento;
 
 @Entity
 @Table(name="user")
-public class User extends Persona{
+public class User {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private int id;
+	
+	@Column(name="nombre")
+	private String nombre;
+	
+	@Column(name="apellido")
+	private String apellido;
+	
+	@Column(name="tipodocumento")
+	@Enumerated(EnumType.ORDINAL)
+	private TipoDocumento tipoDocumento;
+	
+	@Column(name="documento")
+	private long documento;
+	
+	@Column(name="email")
+	private String email;
 	
 	@Column(name="username", unique=true, nullable= false, length=45)
 	private String username;
@@ -25,12 +55,26 @@ public class User extends Persona{
 	@Column(name="activo")
 	private boolean activo;
 	
+	@Column(name="createdat", nullable = false, updatable = false)
+	@CreationTimestamp
+	@ReadOnlyProperty
+	private LocalDateTime createdAt;
+	
+	@Column(name="updatedat")
+	@UpdateTimestamp
+	private LocalDateTime updatedAt;
+	
 	public User() {}
 
 
 	public User(String nombre, String apellido, TipoDocumento tipoDocumento, long documento, String email,
 			String username, String password, UserRole userRole, boolean activo) {
-		super(nombre, apellido, tipoDocumento, documento, email);
+		super();
+		this.nombre = nombre;
+		this.apellido = apellido;
+		this.tipoDocumento = tipoDocumento;
+		this.documento = documento;
+		this.email = email;
 		this.username = username;
 		this.password = password;
 		this.userRole = userRole;
@@ -39,11 +83,66 @@ public class User extends Persona{
 
 	public User(int id, String nombre, String apellido, TipoDocumento tipoDocumento, long documento, String email,
 			String username, String password, UserRole userRole, boolean activo) {
-		super(id, nombre, apellido, tipoDocumento, documento, email);
+		super();
+		this.id = id;
+		this.nombre = nombre;
+		this.apellido = apellido;
+		this.tipoDocumento = tipoDocumento;
+		this.documento = documento;
+		this.email = email;
 		this.username = username;
 		this.password = password;
 		this.userRole = userRole;
 		this.activo = activo;
+	}
+
+	
+	public int getId() {
+		return id;
+	}
+
+	public void setId(int id) {
+		this.id = id;
+	}
+
+	public String getNombre() {
+		return nombre;
+	}
+
+	public void setNombre(String nombre) {
+		this.nombre = nombre;
+	}
+
+	public String getApellido() {
+		return apellido;
+	}
+
+	public void setApellido(String apellido) {
+		this.apellido = apellido;
+	}
+
+	public TipoDocumento getTipoDocumento() {
+		return tipoDocumento;
+	}
+
+	public void setTipoDocumento(TipoDocumento tipoDocumento) {
+		this.tipoDocumento = tipoDocumento;
+	}
+
+	public long getDocumento() {
+		return documento;
+	}
+
+	public void setDocumento(long documento) {
+		this.documento = documento;
+	}
+
+	public String getEmail() {
+		return email;
+	}
+
+	public void setEmail(String email) {
+		this.email = email;
 	}
 
 	public String getUsername() {
@@ -76,6 +175,22 @@ public class User extends Persona{
 
 	public void setActivo(boolean activo) {
 		this.activo = activo;
+	}
+
+	public LocalDateTime getCreatedAt() {
+		return createdAt;
+	}
+
+	public void setCreatedAt(LocalDateTime createdAt) {
+		this.createdAt = createdAt;
+	}
+
+	public LocalDateTime getUpdatedAt() {
+		return updatedAt;
+	}
+
+	public void setUpdatedAt(LocalDateTime updatedAt) {
+		this.updatedAt = updatedAt;
 	}
 	
 }
