+++ src/main/resources/templates/general.html
@@ -13,15 +13,6 @@
                 <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
                     <li class="nav-item"><a class="nav-link active" aria-current="page" th:href="@{/}">Home</a></li>
                     <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">Permisos</a>
-                        <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
-                            <li><a class="dropdown-item" href="#">Solicitar</a></li>
-                            <li><a class="dropdown-item" href="#">Buscar por persona</a></li>
-                            <li><a class="dropdown-item" href="#" sec:authorize="hasRole('ROLE_AUDITOR')">Buscar por rodado</a></li>
-                            <li><a class="dropdown-item" href="#" sec:authorize="hasRole('ROLE_AUDITOR')">Buscar por fecha</a></li>
-                        </ul>
-                    </li>
-                    <li class="nav-item dropdown" sec:authorize="isAuthenticated()">
                         <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">Usuarios</a>
                         <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
                             <li><a class="dropdown-item" th:href="@{/user/abm}" sec:authorize="hasRole('ROLE_ADMIN')">ABM</a></li>
@@ -29,7 +20,7 @@
                             <li><a class="dropdown-item" th:href="@{/user/pdf}" sec:authorize="hasRole('ROLE_AUDITOR')">Generar PDF</a></li>
                         </ul>
                     </li>
-                    <li class="nav-item dropdown" sec:authorize="isAuthenticated()">
+                    <li class="nav-item dropdown">
                         <a class="nav-link dropdown-toggle" id="navbarDropdown" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">Perfiles</a>
                         <ul class="dropdown-menu dropdown-menu-end" aria-labelledby="navbarDropdown">
                             <li><a class="dropdown-item" th:href="@{/userrole/abm}" sec:authorize="hasRole('ROLE_ADMIN')">ABM</a></li>
